+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityConditionComponents.java
@@ -72,8 +72,8 @@ public class EntityConditionComponents implements ConditionComponents {
 	}
 
 	@Override
-	public boolean supports(Class<?> valueClass) {
-		return SUPPORTED_TYPES.contains(requireNonNull(valueClass));
+	public boolean supports(Class<?> type) {
+		return SUPPORTED_TYPES.contains(requireNonNull(type));
 	}
 
 	@Override